requirements.txt
python-telegram-bot
requests
python-dotenv
spipqlite3
anthropic